+++ rusk-uplink/src/lib.rs
@@ -30,3 +30,6 @@ pub use bufwriter::*;
 
 pub mod debug;
 pub use debug::*;
+
+pub mod framing;
+pub use framing::*;
